- split configs.
